refactor : profile service 계층 CQRS 패턴에 메서드 위치 조정 & 테스트 위치 조정
