+++ src/main/java/com/meoguri/linkocean/controller/profile/ProfileController.java
@@ -31,9 +31,10 @@ import com.meoguri.linkocean.domain.bookmark.service.CategoryService;
 import com.meoguri.linkocean.domain.bookmark.service.TagService;
 import com.meoguri.linkocean.domain.profile.persistence.command.dto.ProfileFindCond;
 import com.meoguri.linkocean.domain.profile.service.command.ProfileService;
-import com.meoguri.linkocean.domain.profile.service.command.dto.GetDetailedProfileResult;
-import com.meoguri.linkocean.domain.profile.service.command.dto.GetProfileTagsResult;
-import com.meoguri.linkocean.domain.profile.service.command.dto.GetProfilesResult;
+import com.meoguri.linkocean.domain.profile.service.query.ProfileQueryService;
+import com.meoguri.linkocean.domain.profile.service.query.dto.GetDetailedProfileResult;
+import com.meoguri.linkocean.domain.profile.service.query.dto.GetProfileTagsResult;
+import com.meoguri.linkocean.domain.profile.service.query.dto.GetProfilesResult;
 import com.meoguri.linkocean.infrastructure.s3.S3Uploader;
 
 import lombok.RequiredArgsConstructor;
@@ -48,8 +49,10 @@ public class ProfileController {
 	private static final String PROFILES = "profiles";
 
 	private final ProfileService profileService;
+	private final ProfileQueryService profileQueryService;
 	private final CategoryService categoryService;
 	private final TagService tagService;
+
 	private final S3Uploader s3Uploader;
 
 	/* 프로필 등록 */
@@ -76,7 +79,7 @@ public class ProfileController {
 		final @PathVariable long profileId
 	) {
 		// TODO - 얘 혼자 응답을 말아주는 로직이 컨트롤러에 위치하고 있음 서비스로 옮기던가 나중에 영속성에서 DTO 로 퍼올릴때 수정 할 것
-		final GetDetailedProfileResult profile = profileService.getByProfileId(user.getProfileId(), profileId);
+		final GetDetailedProfileResult profile = profileQueryService.getByProfileId(user.getProfileId(), profileId);
 		final List<GetProfileTagsResult> tags = tagService.getTags(profileId);
 		final List<Category> categories = categoryService.getUsedCategories(profileId);
 
@@ -104,7 +107,7 @@ public class ProfileController {
 		final @RequestParam String username,
 		final Pageable pageable
 	) {
-		final Slice<GetProfilesResult> results = profileService.getProfiles(
+		final Slice<GetProfilesResult> results = profileQueryService.getProfiles(
 			user.getProfileId(),
 			ProfileFindCond.builder()
 				.username(username)
@@ -126,7 +129,7 @@ public class ProfileController {
 		final @PathVariable long profileId,
 		final Pageable pageable
 	) {
-		final Slice<GetProfilesResult> results = profileService.getProfiles(
+		final Slice<GetProfilesResult> results = profileQueryService.getProfiles(
 			user.getProfileId(),
 			ProfileFindCond.builder()
 				.profileId(profileId)
@@ -149,7 +152,7 @@ public class ProfileController {
 		final @PathVariable long profileId,
 		final Pageable pageable
 	) {
-		final Slice<GetProfilesResult> results = profileService.getProfiles(
+		final Slice<GetProfilesResult> results = profileQueryService.getProfiles(
 			user.getProfileId(),
 			ProfileFindCond.builder()
 				.profileId(profileId)
